requirements.txt
Flask>=2.2
Jinja2>=3.1.2
certifi==2019.3.9
chardet==3.0.4
click>=8.1.3
idna==2.8
itsdangerous>=2.2
MarkupSafe>=2.0
python-dotenv==0.10.2
requests==2.22.0
urllib3==1.25.2
Werkzeug>=3.1
